+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 ### Added
 - Added plugin icon.
 
+### Fixed
+- Fixed drag and drop colors.
+
 ## [0.2.0] - 2023-12-13
 ### Changed
 - Updated colors for Sakura Night theme beta.