+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Fixed
 - Fixed drag and drop colors.
+- Fixed editor preview colors.
 
 ## [0.2.0] - 2023-12-13
 ### Changed